+++ src/main/java/com/github/luisflp_dev/rest_escola/academico/instituicao/Instituicao.java
@@ -16,7 +16,7 @@ public class Instituicao {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_instituicao")
-    private Long idInstituicao;
+    private Integer idInstituicao;
 
     @Column(name = "tx_sigla", length = 15,nullable = false, unique = true)
     private String sigla;
