+++ src/main/java/com/github/luisflp_dev/rest_escola/pessoa/aluno/AlunoResponseDTO.java
@@ -5,7 +5,7 @@ import com.github.luisflp_dev.rest_escola.pessoa.Sexo;
 import java.time.LocalDate;
 
 public record AlunoResponseDTO(
-        Long idAluno,
+        Integer idAluno,
         String nome,
         Sexo sexo,
         LocalDate dataNascimento
